+++ src/features/basic/index.ts
@@ -53,6 +53,7 @@ import './mu-fix-sector-names';
 import './nots-material-ticker';
 import './nots-notification-type-label';
 import './nots-ship-arrival-inventory';
+import './other-context-notification-count';
 import './prevent-delete-button-misclicks';
 import './prod-burn-link';
 import './prod-order-eta';